requirements.txt
asgiref==3.8.1
attrs==25.4.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.3
Django==5.0
django-allauth==65.13.1
django-ckeditor-5==0.2.17
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-extensions==4.1
django-filter==25.1
django-lifecycle==1.2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.5.3
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-decouple==3.8
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.4
rpds-py==0.30.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
